requirements.txt
google-adk==1.14.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.10.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
